Update dependencies and switch to go 1.17
Updated to the latest version of the resolution resolver helpers, which
in turn updated knative and some kubernetes libraries to bring them more
in line with what's in pipelines.

A replace directive is added for k8s.io/client-go because pipelines currently
has a misconfigured version in its go.mod. See pull request https://github.com/tektoncd/pipeline/pull/4720/

The parameters for the clusterresolver are also moved to the top of the
readme to make them more apparent.